# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Property rental management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PropertyStatus =
  (typeof PropertyStatus)[keyof typeof PropertyStatus];

export const PropertyStatus = {
  available: "available",
  occupied: "occupied",
  maintenance: "maintenance",
} as const;

export interface Property {
  id: number;
  name: string;
  address: string;
  type: string;
  bedrooms: number;
  bathrooms: number;
  monthlyRent: number;
  status: PropertyStatus;
  imageUrl?: string | null;
  createdAt: string;
}

export type CreatePropertyStatus =
  (typeof CreatePropertyStatus)[keyof typeof CreatePropertyStatus];

export const CreatePropertyStatus = {
  available: "available",
  occupied: "occupied",
  maintenance: "maintenance",
} as const;

export interface CreateProperty {
  name: string;
  address: string;
  type: string;
  bedrooms: number;
  bathrooms: number;
  monthlyRent: number;
  status?: CreatePropertyStatus;
  imageUrl?: string | null;
}

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  confirmed: "confirmed",
  pending: "pending",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface Reservation {
  id: number;
  propertyId: number;
  guestName: string;
  checkIn: string;
  checkOut: string;
  totalAmount: number;
  status: ReservationStatus;
  notes?: string | null;
  createdAt: string;
  propertyName?: string;
}

export type CreateReservationStatus =
  (typeof CreateReservationStatus)[keyof typeof CreateReservationStatus];

export const CreateReservationStatus = {
  confirmed: "confirmed",
  pending: "pending",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface CreateReservation {
  propertyId: number;
  guestName: string;
  checkIn: string;
  checkOut: string;
  totalAmount: number;
  status?: CreateReservationStatus;
  notes?: string | null;
}

export interface Expense {
  id: number;
  propertyId: number;
  category: string;
  description: string;
  amount: number;
  date: string;
  createdAt: string;
  propertyName?: string;
}

export interface CreateExpense {
  propertyId: number;
  category: string;
  description: string;
  amount: number;
  date: string;
}

export type ListReservationsParams = {
  propertyId?: number;
};

export type ListExpensesParams = {
  propertyId?: number;
};