# lib/api-zod/src/generated/types/createProperty.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Property rental management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePropertyStatus } from "./createPropertyStatus";

export interface CreateProperty {
  name: string;
  address: string;
  type: string;
  bedrooms: number;
  bathrooms: number;
  monthlyRent: number;
  status?: CreatePropertyStatus;
  imageUrl?: string | null;
}