# lib/api-zod/src/generated/types/createReservation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Property rental management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateReservationStatus } from "./createReservationStatus";

export interface CreateReservation {
  propertyId: number;
  guestName: string;
  checkIn: Date;
  checkOut: Date;
  totalAmount: number;
  status?: CreateReservationStatus;
  notes?: string | null;
}