# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Property rental management API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  propertyId: number;
  category: string;
  description: string;
  amount: number;
  date: Date;
  createdAt: Date;
  propertyName?: string;
}