# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Property rental management API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyStatus } from "./propertyStatus";

export interface Property {
  id: number;
  name: string;
  address: string;
  type: string;
  bedrooms: number;
  bathrooms: number;
  monthlyRent: number;
  status: PropertyStatus;
  imageUrl?: string | null;
  createdAt: Date;
}