# lib/api-zod/src/generated/types/reservation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Property rental management API
 * OpenAPI spec version: 0.1.0
 */
import type { ReservationStatus } from "./reservationStatus";

export interface Reservation {
  id: number;
  propertyId: number;
  guestName: string;
  checkIn: Date;
  checkOut: Date;
  totalAmount: number;
  status: ReservationStatus;
  notes?: string | null;
  createdAt: Date;
  propertyName?: string;
}